Move subreddit stats to separate file
