+++ src/analizer.ts
@@ -1,6 +1,7 @@
 import { Comment } from 'snoowrap';
 import Logger from './logger';
 import { getAllUserComments, isToRMod } from './reddit-api';
+import { analyzeSubreddits } from './stats/subreddits';
 import { CountTag, specialTags, countTags, Tag } from './tags';
 import Transcription from './transcription';
 import { limitEnd } from './util';
@@ -193,43 +194,6 @@ export function getTranscriptionAmount(transcriptions: Transcription[]): Transcr
     wordAvg: wordTotal / count,
     wordPeak,
   };
-}
-
-type SubStats = {
-  /** The name of the subreddit. */
-  sub: string;
-  /** The number of transcriptions for the subreddit. */
-  count: number;
-};
-
-/**
- * Analyzes the subreddits of the transcriptions.
- * @param transcriptions The transcriptions to analyze.
- */
-export function analyzeSubreddits(transcriptions: Transcription[]): SubStats[] {
-  const subStats: SubStats[] = [];
-
-  transcriptions.forEach((transcription) => {
-    const sub = transcription.subredditNamePrefixed;
-
-    const stats = subStats.find((stat) => {
-      return stat.sub === sub;
-    });
-
-    if (stats) {
-      stats.count += 1;
-    } else {
-      subStats.push({
-        sub,
-        count: 1,
-      });
-    }
-  });
-
-  // Sort by count descending
-  return subStats.sort((a, b) => {
-    return b.count - a.count;
-  });
 }
 
 type FormatStats = {
